pyproject.toml
[project]
name = "premortem"
version = "0.1.0"
description = "LLM-driven failure analysis for distributed system designs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.scripts]
premortem = "premortem.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"
